fix(backend): change subject credit type to double precision
- Changed 'credit' column from DECIMAL(3,1) to DOUBLE PRECISION to match Rust's f64 type to prevent sqlx decoding errors

+++ backend-school/migrations/023_create_curriculum_basics.sql
@@ -42,7 +42,7 @@ CREATE TABLE IF NOT EXISTS subjects (
     name_en VARCHAR(200),
     
     -- Properties
-    credit DECIMAL(3,1) NOT NULL DEFAULT 0.0,
+    credit DOUBLE PRECISION NOT NULL DEFAULT 0.0,
     hours_per_semester INTEGER,            -- 40, 60, 80
     
     -- Classification
